Factor consensus engine and reward logic out
